moved data model classes to org.stellar.anchor.platform.model package
