+++ platform/src/main/java/org/stellar/anchor/platform/data/PaymentStreamerCursorRepo.java
@@ -3,6 +3,7 @@ package org.stellar.anchor.platform.data;
 import java.util.Optional;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.data.repository.CrudRepository;
+import org.stellar.anchor.platform.model.PaymentStreamerCursor;
 
 public interface PaymentStreamerCursorRepo extends CrudRepository<PaymentStreamerCursor, String> {
