+++ platform/src/main/java/org/stellar/anchor/platform/model/PaymentStreamerCursor.java
@@ -1,4 +1,4 @@
-package org.stellar.anchor.platform.data;
+package org.stellar.anchor.platform.model;
 
 import com.google.gson.annotations.SerializedName;
 import javax.persistence.*;
@@ -7,9 +7,9 @@ import lombok.Data;
 @Data
 @Entity
 @Access(AccessType.FIELD)
-@Table(name = "stellar_account_page_token")
+@Table(name = "stellar_payment_observer_page_token")
 public class PaymentStreamerCursor {
-  public static final String SINGLETON_ID = "ALL";
+  public static final String SINGLETON_ID = "SINGLETON_ID";
 
   @Id
   @SerializedName("id")
