+++ platform/src/main/java/org/stellar/anchor/platform/model/SepTransaction.java
@@ -1,4 +1,4 @@
-package org.stellar.anchor.platform.data;
+package org.stellar.anchor.platform.model;
 
 /** This is the interface to tag SEP transaction classes. */
 public interface SepTransaction {}
